feat: consolidate GPU test utilities and enhance Edge tests with GPU support
